test time

+++ src/main/java/com/netcracker/dao/impl/mapper/CreditDebtMapper.java
@@ -16,7 +16,7 @@ public class CreditDebtMapper implements RowMapper<Debt> {
                 .debtId(resultSet.getBigDecimal("debt_id").toBigInteger())
                 .dateFrom(checkNullDate(resultSet.getDate("debt_from")))
                 .dateTo(checkNullDate(resultSet.getDate("debt_to")))
-                .amountDebt(resultSet.getLong("debt_amount"))
+                .amountDebt(resultSet.getDouble("debt_amount"))
                 .build();
     }
